common func was changed
